# docs/artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mini-conf API",
    "description": "HTTP-first deployment configuration APIs for mini-conf",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/open/deployments/{deployment_key}/config-bundle": {
      "get": {
        "tags": [
          "open"
        ],
        "operationId": "get_config_bundle",
        "parameters": [
          {
            "name": "deployment_key",
            "in": "path",
            "description": "Deployment instance key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current config bundle for a deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigBundleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path or query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Bearer token does not match target deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Deployment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Database bootstrap disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/open/heartbeats": {
      "post": {
        "tags": [
          "open"
        ],
        "operationId": "report_heartbeat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Heartbeat accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentSyncResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Bearer token does not match target deployment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Deployment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Database bootstrap disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigBundleItem": {
        "type": "object",
        "required": [
          "config",
          "revision",
          "content_hash",
          "format",
          "content"
        ],
        "properties": {
          "config": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "content_hash": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "revision": {
            "type": "string"
          }
        }
      },
...
      "ConfigBundleResponse": {
        "type": "object",
        "required": [
          "project",
          "environment",
          "deployment",
          "configs"
        ],
        "properties": {
          "configs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigBundleItem"
            }
          },
          "deployment": {
            "$ref": "#/components/schemas/ResolveDeployment"
          },
          "environment": {
            "type": "string"
          },
          "project": {
            "type": "string"
          }
        }
      },
...
      "DeploymentSyncResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "HeartbeatRequestBody": {
        "type": "object",
        "required": [
          "project",
          "environment",
          "deployment_key",
          "process_key",
          "metadata"
        ],
        "properties": {
          "deployment_key": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "metadata": {
            "type": [
              "object",
              "null"
            ]
          },
          "process_key": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "reported_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ResolveDeployment": {
        "type": "object",
        "required": [
          "key",
          "name"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      },
      "session_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mini_conf_session"
      }
    }
  },
...
}